Fix: Meal-94-BE-해당날짜의끼니기록반환api 반환 값 수정 (meal_kcal: double->int)

+++ src/main/java/mealplanb/server/service/MealService.java
@@ -78,7 +78,7 @@ public class MealService {
         log.info("[MealService.makeGetMealItems]");
         List<GetMealItem> mealItems = new ArrayList<>();
         for (Meal meal : meals){
-            GetMealItem getMealItem = new GetMealItem(meal.getMealId() ,MealTypeConverter.convertMealTypeLabel(meal.getMealType()), foodMealMappingTableService.getMealKcal(meal.getMealId()));
+            GetMealItem getMealItem = new GetMealItem(meal.getMealId() ,MealTypeConverter.convertMealTypeLabel(meal.getMealType()), (int) foodMealMappingTableService.getMealKcal(meal.getMealId()));
             mealItems.add(getMealItem);
         }
         return mealItems;
